M3 findings, M5 decision, docs, clippy clean
specs/m3-findings.md records the budget-gate outcome (small passes; four
representative-scale overruns explained as hub tails, unbounded
shortest_path, decay-batch modeling, and confirmed filtered-similarity
posture) — the §16.3 "written reason to revise" M3 exit.

specs/m5-query-seam-decision.md records the M5 decision: ship the
PropertyGraphRead seam, defer the Cypher/GQL binding to v0.2.

CLAUDE.md updated from pre-code to the v0.1 layout, with build/bench
commands. Clippy clean across the workspace.

+++ harness/src/generator.rs
@@ -140,7 +140,7 @@ fn p_cat(within: u64, pop: u64) -> String {
         "cat_common".into()
     } else {
         // Filler categories, ~1% each, so none rivals the distinguished bands.
-        let filler = (within % 100) as u64;
+        let filler = within % 100;
         format!("cat_f{filler:02}")
     }
 }
